+++ tf/modules/k8s-secrets/main.tf
@@ -1,14 +1,3 @@
-resource "kubernetes_secret" "mailgun-api-key" {
-  metadata {
-    name = "mailgun-api-key"
-    namespace = "default"
-  }
-
-  data = {
-    "key" = var.domain_mailgun_key
-  }
-}
-
 resource "kubernetes_secret" "smtp-credentials" {
   metadata {
     name = "smtp-credentials"